requirements.txt
pymongo==4.6.3
python-dotenv==1.0.0